5.1.14 Создаем класс BoardController, и в него выносим часть связанной логики.
